put BootstrapDialog into a ui folder
