+++ frontend/src/components/UI/BootstrapDialog.ts
@@ -1,0 +1,13 @@
+import { styled } from '@mui/material/styles';
+import Dialog from '@mui/material/Dialog';
+
+const BootstrapDialog = styled(Dialog)(({ theme }) => ({
+    '& .MuiDialogContent-root': {
+        padding: theme.spacing(2),
+    },
+    '& .MuiDialogActions-root': {
+        padding: theme.spacing(1),
+    },
+  }));
+
+export default BootstrapDialog;
